requirements.txt
streamlit
psycopg2-binary
python-dotenv
panda
SQLite
pip install psycopg2
psycopg2
